API: IsNullAt enhanced

+++ Polars.CSharp.Tests/GroupByTests.cs
@@ -11,8 +11,8 @@ public class TimeSeriesTests
     {
         // 10:00, 10:10, 10:20, 10:30, 10:40, 10:50
         var start = new DateTime(2024, 1, 1, 10, 0, 0);
-        var end = new DateTime(2024, 1, 1, 10, 50, 0); 
-        var values = new[] { 0, 1, 2, 3, 4, 5 }; 
+        var end = new DateTime(2024, 1, 1, 10, 50, 0);
+        var values = new[] { 0, 1, 2, 3, 4, 5 };
 
         var df = DataFrame.FromColumns(new { Val = values })
             .WithColumns(
@@ -33,10 +33,10 @@ public class TimeSeriesTests
             );
 
         Assert.Equal(2, res.Height);
-        
+
         Assert.Equal(3, res.GetValue<int>(0, "SumVal"));
         Assert.Equal(3, res.GetValue<int>(0, "Count"));
-        
+
         Assert.Equal(12, res.GetValue<int>(1, "SumVal"));
         Assert.Equal(3, res.GetValue<int>(1, "Count"));
     }
@@ -50,30 +50,30 @@ public class TimeSeriesTests
 
         var df = DataFrame.FromColumns(new {Time = dates,Val = values});
 
-        // Every: 5m 
-        // Period: 10m 
+        // Every: 5m
+        // Period: 10m
         // Label: Right
         // IncludeBoundaries: True
         var res = df
             .GroupByDynamic(
                 indexColumn: Cs.Datetime(),
-                every: TimeSpan.FromMinutes(5),  
+                every: TimeSpan.FromMinutes(5),
                 period: TimeSpan.FromMinutes(10),
-                label: Label.Right,             
-                includeBoundaries: true,         
-                ClosedInterval: ClosedInterval.Left  
+                label: Label.Right,
+                includeBoundaries: true,
+                ClosedInterval: ClosedInterval.Left
             )
             .Agg(
                 Pl.Col("Val").Count().Alias("Count")
             );
 
-        // Window 1: [09:55, 10:05) -> Label 10:05. 
-        // Window 2: [10:00, 10:10) -> Label 10:10. 
-        
+        // Window 1: [09:55, 10:05) -> Label 10:05.
+        // Window 2: [10:00, 10:10) -> Label 10:10.
+
         Assert.Contains("_lower_boundary", res.ColumnNames);
         Assert.Contains("_upper_boundary", res.ColumnNames);
-        
-        var firstTime = res.GetValue<DateTime>(0, "Time"); 
+
+        var firstTime = res.GetValue<DateTime>(0, "Time");
         Assert.True(res.Height > 0);
     }
     [Fact]
@@ -81,10 +81,10 @@ public class TimeSeriesTests
     public void Test_GroupByDynamic_Having_And_MultipleAggregations()
     {
         var start = new DateTime(2024, 1, 1, 10, 0, 0);
-        var end = new DateTime(2024, 1, 1, 11, 20, 0); 
-        
+        var end = new DateTime(2024, 1, 1, 11, 20, 0);
+
         // Group 1 [10:00, 10:30): 10, 15, 5 -> Sum: 30, Max: 15, Min: 5
-        // Group 2 [10:30, 11:00):  2,  3, 4 -> Sum: 9,  Max: 4,  Min: 2  
+        // Group 2 [10:30, 11:00):  2,  3, 4 -> Sum: 9,  Max: 4,  Min: 2
         // Group 3 [11:00, 11:30): 20, 25, 5 -> Sum: 50, Max: 25, Min: 5
         int[] values = [10, 15, 5, 2, 3, 4, 20, 25, 5];
 
@@ -99,7 +99,7 @@ public class TimeSeriesTests
                 every: TimeSpan.FromMinutes(30),
                 ClosedInterval: ClosedInterval.Left
             )
-            .Having(Pl.Col("Val").Sum() > 20) 
+            .Having(Pl.Col("Val").Sum() > 20)
             .Agg(
                 Pl.Col("Val").Sum().Alias("SumVal"),
                 Pl.Col("Val").Max().Alias("MaxVal"),
@@ -111,7 +111,7 @@ public class TimeSeriesTests
         Assert.Equal(30, res.GetValue<int>(0, "SumVal"));
         Assert.Equal(15, res.GetValue<int>(0, "MaxVal"));
         Assert.Equal(5, res.GetValue<int>(0, "MinVal"));
-        
+
         Assert.Equal(50, res.GetValue<int>(1, "SumVal"));
         Assert.Equal(25, res.GetValue<int>(1, "MaxVal"));
         Assert.Equal(5, res.GetValue<int>(1, "MinVal"));
@@ -122,10 +122,10 @@ public class TimeSeriesTests
     public void Test_GroupByDynamic_With_By_Column_And_Having()
     {
         var start = new DateTime(2024, 1, 1, 10, 0, 0);
-        var dates = new[] 
+        var dates = new[]
         {
-            start, start.AddMinutes(30), 
-            start, start.AddMinutes(30) 
+            start, start.AddMinutes(30),
+            start, start.AddMinutes(30)
         };
         var symbols = new[] { "A", "A", "B", "B" };
         var values = new[] { 10, 20, 100, 200 };
@@ -135,19 +135,19 @@ public class TimeSeriesTests
         var res = df
             .GroupByDynamic(
                 indexColumn: "Time",
-                every: "1h", 
+                every: "1h",
                 groupBy: [Pl.Col("Symbol")]
             )
-            .Having(Pl.Col("Val").Mean() > 50) 
+            .Having(Pl.Col("Val").Mean() > 50)
             .Agg(
                 Pl.Col("Val").Mean().Alias("MeanVal"),
                 Pl.Col("Val").First().Alias("FirstVal"),
                 Pl.Col("Val").Last().Alias("LastVal")
             );
         Assert.Equal(1, res.Height);
-        
+
         Assert.Equal("B", res.GetValue<string>(0, "Symbol"));
-        Assert.Equal(150.0, res.GetValue<double>(0, "MeanVal")); 
+        Assert.Equal(150.0, res.GetValue<double>(0, "MeanVal"));
         Assert.Equal(100, res.GetValue<int>(0, "FirstVal"));
         Assert.Equal(200, res.GetValue<int>(0, "LastVal"));
     }
@@ -157,15 +157,15 @@ public class TimeSeriesTests
     {
         var start = new DateTime(2024, 1, 1).Ticks;
         var dates = Enumerable.Range(0, 100)
-            .Select(i => new DateTime(start + i)) 
+            .Select(i => new DateTime(start + i))
             .ToArray();
-        
-        using var df = DataFrame.FromColumns(new { 
+
+        using var df = DataFrame.FromColumns(new {
             Ts = dates,       // DateTime[]
             Val = dates       // DateTime[]
         });
 
-        var us1 = TimeSpan.FromTicks(10); 
+        var us1 = TimeSpan.FromTicks(10);
 
         using var res = df
             .GroupByDynamic(
@@ -177,7 +177,7 @@ public class TimeSeriesTests
             );
 
         Assert.Equal(10, res.Height);
-        
+
         Assert.Equal(10, res.GetValue<int>(0, "Count"));
     }
     [Fact]
@@ -198,7 +198,7 @@ public class TimeSeriesTests
         Assert.Equal(new DateTime(2024, 1, 1, 10, 30, 0), res["dt_ms"][0]);
         Assert.Equal(new DateTime(2024, 1, 1, 10, 30, 0),res["dt_us"][0]);
 
-        Assert.Equal(new DateTime(2024, 12, 31, 23, 59, 59, 123),res["dt_ms"][1]); 
+        Assert.Equal(new DateTime(2024, 12, 31, 23, 59, 59, 123),res["dt_ms"][1]);
         Assert.Equal(new DateTime(2024, 12, 31, 23, 59, 59, 123), res["dt_us"][1]);
 
         Assert.Equal(DataTypeKind.Datetime, res.Schema["dt_ms"].Kind);
@@ -219,7 +219,7 @@ public class TimeSeriesTests
             .Rolling(
                 indexColumn: "Time",
                 period: "20m",
-                ClosedInterval: ClosedInterval.Both 
+                ClosedInterval: ClosedInterval.Both
             )
             .Agg(
                 Pl.Col("Val").Sum().Alias("SumVal"),
@@ -228,13 +228,13 @@ public class TimeSeriesTests
         Assert.Equal(5, res.Height);
 
         Assert.Equal(10, res.GetValue<int>(0, "SumVal"));
-        Assert.Equal(1, res.GetValue<int>(0, "Count"));
+        Assert.Equal(1, res.GetValue<long>(0, "Count"));
 
         Assert.Equal(30, res.GetValue<int>(1, "SumVal"));
-        Assert.Equal(2, res.GetValue<int>(1, "Count"));
+        Assert.Equal(2, res.GetValue<long>(1, "Count"));
 
         Assert.Equal(60, res.GetValue<int>(2, "SumVal"));
-        Assert.Equal(3, res.GetValue<int>(2, "Count"));
+        Assert.Equal(3, res.GetValue<long>(2, "Count"));
     }
 
     [Fact]
@@ -254,10 +254,10 @@ public class TimeSeriesTests
 
         var res = df
             .Rolling(
-                indexColumn: Cs.Temporal(), 
+                indexColumn: Cs.Temporal(),
                 period: TimeSpan.FromMinutes(30),
-                groupBy: ["Symbol"], 
-                ClosedInterval: ClosedInterval.Right 
+                groupBy: ["Symbol"],
+                ClosedInterval: ClosedInterval.Right
             )
             .Agg(
                 Pl.Col("Val").Max().Alias("MaxVal")
@@ -276,7 +276,7 @@ public class TimeSeriesTests
     public void Test_GroupByDynamic_Head_And_Tail()
     {
         var start = new DateTime(2024, 1, 1, 10, 0, 0);
-        var end = new DateTime(2024, 1, 1, 10, 50, 0); 
+        var end = new DateTime(2024, 1, 1, 10, 50, 0);
         var values = new[] { 0, 1, 2, 3, 4, 5 };
 
         var df = DataFrame.FromColumns(new { Val = values })
@@ -284,13 +284,13 @@ public class TimeSeriesTests
 
         // Window 1 [10:00, 10:30): 10:00(0), 10:10(1), 10:20(2)
         // Window 2 [10:30, 11:00): 10:30(3), 10:40(4), 10:50(5)
-        
+
         var headRes = df
             .GroupByDynamic("Time", "30m", ClosedInterval: ClosedInterval.Left)
             .Head(2);
 
         Assert.Equal(4, headRes.Height);
-        
+
         Assert.Equal(0, headRes.GetValue<int>(0, "Val"));
         Assert.Equal(1, headRes.GetValue<int>(1, "Val"));
 
@@ -302,8 +302,8 @@ public class TimeSeriesTests
             .Tail(1);
 
         Assert.Equal(2, tailRes.Height);
-        Assert.Equal(2, tailRes.GetValue<int>(0, "Val")); 
-        Assert.Equal(5, tailRes.GetValue<int>(1, "Val")); 
+        Assert.Equal(2, tailRes.GetValue<int>(0, "Val"));
+        Assert.Equal(5, tailRes.GetValue<int>(1, "Val"));
     }
     public struct DeptKey
     {
@@ -332,7 +332,7 @@ public class TimeSeriesTests
 
         var (Key, Group) = groups.First(g => g.Key.Department == "IT");
         Assert.NotNull(Group);
-        Assert.Equal(3, Group.Height); 
+        Assert.Equal(3, Group.Height);
 
         var ids = Group["Id"].ToArray<int>();
         Assert.True(ids.SequenceEqual([1, 3, 6]));
@@ -366,13 +366,13 @@ public class TimeSeriesTests
             {
                 Assert.Equal(3, groupDf.Height);
             }
-            
-            groupDf.Dispose(); 
+
+            groupDf.Dispose();
         }
 
         Assert.Equal(2, groupCount);
     }
-    
+
     [Fact]
     public void Test_GroupBy_With_MultiKeys()
     {
@@ -386,11 +386,11 @@ public class TimeSeriesTests
         var groups = df.GroupBy(["Department", "Gender"]).GetGroups<MultiKey>().ToList();
 
         Assert.Equal(2, groups.Count);
-        
+
         var (Key, Group) = groups.FirstOrDefault(g => g.Key.Department == "IT" && g.Key.Gender == "M");
         Assert.NotNull(Group);
         Assert.Equal(2, Group.Height);
-        
+
         foreach (var g in groups) { g.Group.Dispose(); }
     }
 }
